# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExportKit API",
    "version": "1.1.0",
    "description": "Production-ready data export API for SaaS teams. Generate CSV, JSON, and Excel exports with webhooks, signed URLs, email notifications, billing, team management, and scheduling.",
    "contact": {
      "name": "ExportKit Support",
      "email": "support@exportkit.io",
      "url": "https://exportkit.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.exportkit.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/webhooks/stripe": {
      "post": {
        "summary": "Stripe webhook handler",
        "description": "Handles incoming Stripe webhook events for subscription lifecycle management. Requires a valid `stripe-signature` header for verification. This endpoint uses raw body parsing.",
        "tags": ["Webhooks"],
        "parameters": [
          {
            "in": "header",
            "name": "stripe-signature",
            "required": true,
            "schema": { "type": "string" },
            "description": "Stripe webhook signature for event verification"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Raw Stripe event payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": { "type": "boolean" }
                  }
                },
                "example": { "received": true }
              }
            }
          },
          "400": {
            "description": "Invalid signature or webhook processing error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Missing stripe-signature header" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" },
          "code": { "type": "string", "description": "Machine-readable error code (UPPER_SNAKE_CASE)" },
          "details": {
            "type": "array",
            "items": { "type": "object" },
            "description": "Validation error details (when applicable)"
          }
        },
        "required": ["error", "code"]
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Generate keys via the Dashboard or POST /api/keys endpoint."
      }
    }
  }
}